moved cpu_state in another file, deserialize and compare cycles
