+++ aliusnes/tests/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod cpu_state;
 pub mod test_bus;
